// ssfi-frontend/src/types/cms.ts
export interface SocialLinks {
  facebook?: string;
  twitter?: string;
  instagram?: string;
  youtube?: string;
  linkedin?: string;
}

export interface SiteSettings {
  id?: string;
  siteName?: string;
  siteTagline?: string;
  logo?: string;
  favicon?: string;
  contactEmail?: string;
  contactPhone?: string;
  address?: string;
  socialLinks?: SocialLinks;
  footerText?: string;
  googleAnalyticsId?: string;
  maintenanceMode?: boolean;
  maintenanceMessage?: string;
}